Exception: Make base methods requisite
The rationale here is that whatever calls into Rogue may want to
be able to handle its exceptions, and doing something useful may
require getting a decent message out of them.

Make the init methods requisite too for good measure (so that a
C++ caller can generate Rogue exceptions which *might* be useful
but should be pretty cheap anyway).