Cython: Implement __copy__/__deepcopy__
Well, sort of implement them. We don't have a general way to copy
or deep copy Rogue objects. We could allow for some way to do it
(e.g., a Rogue _copy_ or whatever), but we don't currently. Instead,
we just return self.  This at least keeps things from breaking.
Without it, I think deep copies were copying the the shadow objects
without incrementing the refcount, which is not good.