Add alias parameters
Adding a "@" to the end of a type in a parameter list now makes it
an alias parameter, which is akin to a C++ reference.  This means
that whatever you pass in for the parameter must be something like
a variable.

A limitation of the current implementation is that references
and array elements cannot be aliased in the "auto" GC mode.

This commit also ends support for changing GC modes by editing the
output C++; you now need to recompile the Rogue.