Improve return values for MethodInfo invocation
Previously, when invoking a method via MethodInfo, you were required to
use the .returns<<...>> template method to specify the return type.  Now
you can use the '->' operator, which is consistent with, e.g.,
native(...)->type.  Internally, it does this by just translating to
the .returns method.

Additionally, if you try to use the return value of a .call invocation,
the compiler gives a special error message, which is more meaningful
than what you'd get without the special handling.

Finally, the return types are now always checked at runtime (we could
put them in an $if so that they are only on in debug mode if anyone
really cares about the performance), and pretty helpful error messages
are generated on failure.