Use C linkage when possible
This changes a lot of stuff to use 'extern "C"'.  This includes methods
which don't use compounds and a bunch of Rogue internal API functions.
If I remember right, the compound limitation is because in auto GC mode,
compounds aren't POD types if they hold references, so aren't technically
allowed to use C linkage (meaning we could probably allow C linkage for
compounds with no references).

The best thing about this is that it exports unmangled names for these
functions.