Add Python/ctypes plugin
Like the Cython plugin, this is an additional "add-on" target for
roguec which generates Python binding code.  Indeed, the code for
it is adapted from the Cython plugin, and the two have much in
common.  *However*, rather than generate Cython code which then
needs to be compiled to C++ and then that C++ needing to be
compiled and linked with the output from the Rogue compiler,
this plugin modifies the output C++ a bit -- which should then
be compiled as a shared library -- and generates a plain
Python file which uses ctypes to load and interact with the
that shared library.

I haven't done any performance benchmarks, but this may be slower
than the Cython approach (though I wouldn't be surprised if it
were not too much slower).  The primary benefit here is that
for a large project, compiling the Cython and the output of the
Cython could be incredibly expensive -- in one case, the
11 megabytes of Cython generated around 170 megabytes of
C++ which took several minutes and several gigabytes of RAM to
compile in a giant 130 megabyte shared library.  The ctypes
approach is much lighter weight.  As a side benefit, you don't
need the Cython compiler, and the build process is a bit
simpler.

The Python plugin makes several optimizations to reduce code
size (factoring out common stuff from generated code, collapsing
code that ends up being identical, etc.), and some of these
could probably be applied to the Cython plugin too.  However,
it seems entirely possible that the Cython plugin will be
phased out in favor of this one.

This isn't incredibly well tested yet.