Preprocessor: More powerful $if/$elseIf expressions
$if/$elseIf are no longer limited strictly to Logical expressions.  This
lets you, e.g., compare string or numeric values.  So we've added some
comparison operators to let you actually do it.

This is a semi-breaking changing.  You should now definitely use
defined() (or the new optional()) when checking defines which may not
be defined.  And the old $if "C++" should now be $if target("C++").