mtgc: Don't deadlock in on_cleanup()
on_cleanup() functions run from the GC thread, but they may also
check to see if the GC wants to run... which is bad, because that
would cause the GC thread to deadlock waiting for itself.

The GC thread now knows that it's the GC thread so it can avoid
waiting for itself.