py: Delegate/thunk fix
This has two tweaks.  The first one is almost surely a bug fix.  In the
generated ctypes thunk code, retval was being adjusted if it needed
to be transformed somehow (e.g., inserting a cast), but then the same
code was being generated again, so it was essentially applying the
transformation twice.

The second part is a change to the semantics of cast_object, so that it
now doesn't return the thisptr of objects, but thisptr.value.  This
doesn't seem great, but I think the alternative is that various
cast-sites become more complex to deal with pointer versus non-pointer
cases?  We'll see how this works.