[String] Added String.replacing_pattern(srch:String,repl:String)->String
Search pattern has same "$*?" wildcard format as .extract_strings():
  - '$' acts as '*' but also denotes a piece to be extracted.
Replacement pattern contains '$' to denote implicitly-paired points of substitution
or'$(0)' etc. to indicate the 0th, etc. extracted value. '$()' can be used to
escape the special meaning of a successive '(0)' and '$($)' can be written to
indicate a literal '$'.

If this string contains line breaks then the search is performed on separate lines
in order to avoid excessive search times by wildcards.

Examples:
  "Alice and Bob".replacing_pattern( "$ and $", "$ & $" ) -> "Alice & Bob"
  "Alice and Bob".replacing_pattern( "$ and $", "$(1) and $(0)" ) -> "Bob and Alice"

- Fixed pattern-matching bugs in .replacing_pattern() and .locate_pattern()
  - .replacing_pattern() now replaces multiple per line
  - pattern search no longer consumes empty spaces in the beginning