python: Improved overload resolution
Previously, any argument that was an instance of a the parameter type
matched perfectly.  But some matches are better than other.  For
example, if a parameter type was Object, it would match any instance
of any class, but *any* other compatible reference parameter type would
be a closer match and therefore preferable.

So now, if we end up with any ties, we go back and re-rank the ties
carefully -- factoring in the "distance" between the argument type
and the parameter type, so that an exact match scores highly, and
every other additional superclass between the argument and the
parameter makes it match a little worse.